Fix FlexSPI LUT access
Fix the SPI dual output mode. There was a bug, where the output lines
were tri-state during the mode cycle, which left the flash chip in read
mode on some circumstances. Fix that by driving the lines low during
that cycle.
Additionally, there was a bug in the unlock code (which didn't affect
anything, because the LUTs were not locked). Since this commit, we also
lock the LUTs after we've written them.

Signed-off-by: Michael Walle <michael.walle@kontron.com>